Auto-quote the prefix variable if needed (#1)
According to the original pkg-config semantics, every .pc file should be
quoting its uses of ${prefix} since it may contain spaces. Literally
nobody does this, so instead automatically quote it if the prefix
contains a space. Unusual, but it *does* come up from time to time on
Windows, even still today.

Note: This will only help if when using eval instead of just command
substitution, which cannot solve this problem.

    $ eval cc app.c $(pkg-config sdl2 --cflags --libs)