Define Size_MASK for hash truncation
MSVC's truncation run-time check (/RTCc) does not accept explicit casts,
and the value must actually be truncated through bit-wise or arithmetic
operations. So define Size_MASK as an unsigned integer of equal width to
a Size. This truncation will be optimized out in real builds but permtis
testing with /RTCc.