Extend system include/lib to other platforms
Arguments are removed via dumb string comparisons that don't try to
understand paths (accounting relative components, etc.). Normally not a
big deal, but the Windows situation is messier than usual:

* Two interchangeable directory separators
* Case insensitivity

It's common for programs to receive equivalent paths that are not
strictly equivalent strings, even when case folding. For now just accept
"system" paths may be missed becauase, say, the drive letter was
capitalized or one of the path configurations used a backslash.