Process system paths like outputs before caching
In particular, if the system path contains a space then it should be
escaped before inserting into the "seen" set. Otherwise it will not
match in the lookup.

The change includes a test of such a practical case on Windows, as well
as the typical non-case on Linux to make sure that didn't break.