Convert source to a C and C++ polyglot
Only applies to the following platforms: generic, win32, and test. The
rest don't need it. The purpose is to enable compilation on ancient
toolchains as C++ source. The tests now run successfully on Windows 95
using MSVC 6.

C++20 introduced a "requires" keyword which clashes with this program,
so it is not valid C++20. However, every existing C++20 compiler will
have no trouble compiling the source as C anyway.