Keep last "-l" instead of first when deduping
In practice this issue only matters for static libraries and Windows
import libraries, which behave like static libraries. In the typical
case where this matters, the problematic import library belongs to a
system DLL.

GCC resolves symbols in one pass, left to right. It tracks undefined
symbols and resolves them with definitions in later inputs. Inputs not
referenced by an undefined symbol may be discarded. Listing an input too
early is like not listing it at all. Thus definitions should follow, not
precede, the inputs requiring them.

Consider package A with libs "-lA -lC" (A needs definitions from C) and
package B with libs "-lB -lC" (B needs definitions from C). There are
three obvious outputs when configuring "pkg-config --libs A B":

Best : -lA -lB -lC
Okay : -lA -lC -lB -lC
Worst: -lA -lC -lB

The worst case nearly always works out because either some inputs are
shared libraries or the main program references A which references C,
keeping it "alive" for B. However, if the main program doesn't reference
A, perhaps because it's an optional feature guarded by an #ifdef, then A
is discarded, as is C, leaving B with undefined symbols. It's similar if
A doesn't reference C strongly enough.

Solution: When removing redundant -l arguments, keep the last instance,
not the first.

This commit only covers the simpler, common case. If C is a package and
A and B do not use "-lC" directly but through "Requires:", then "-lC"
will still not list last. The processing in this commit happens after
the dependency tree is resolved, but moving C to the end requires an
earlier manipulation of the dependency tree.