Use volatile and a memory clobber for strlen()
Since strlen is only used on static strings (getenv, argv), and never on
modified buffers, these make no difference. However, it may be necessary
in other contexts, and it would be a good idea to get it right in case
of copy-pasting.

The volatile prevents the compiler from optimizing away strlen() calls
it believes will return the same result. These guesses are sometimes
wrong. The memory clobber forces register flushes and such before the
call, even when not optimized out.