Support ancient Visual C++ builds
Visual C++ 6.0 and earlier was a challenge because it does not support
any particular C++ standard. It's instead a unique pseudo-C++ dialect.
This change allows compiling at least as far back as Visual C++ 4.2, and
these builds work on Windows NT 3.1 (1993). That makes u-config portable
across 30 years of operating systems. As for the old Visual C++ dialect:

* For loop variables aren't scoped to the loop
* Initializer lists must be constant, like in C89
* Every non-void function must end with a return, no exceptions
* No long long type (see __int64)
* No wchar_t type (see /Zc:wchar_t)

Without Service Pack 6, Visual C++ 6.0 miscompiles getargopt(), which
was fun to debug. The changes to that function are a workaround, though,
there may be other such miscompilations. This one happened to be caught
by the tests.