Process paths with special internal encoding
Instead of implicitly wrapping ${prefix} in quotes, which works in most
cases but breaks in uncommon cases, encode paths with a UTF-8 extension.
The encoding converts white space to UTF-8-illegal bytes which will not
be split during processing. Such bytes cannot appear in valid input, so
it's unambiguous and reversible. The UTF-8-illegal bytes are decoded at
the last moment when printing output.