Extend path encoding to include "$()"
More special characters can be handled using the same technique as white
space. It allows .pc files to emit meta characters literally, but escape
them as path components, which is a somewhat common situation on Windows
("Program Files (x86)"). This includes ${prefix}, ${pcfiledir}, and even
${pc_top_builddir}.

The default ${pc_top_builddir}, "$(top_builddir)", is not path-encoded
because it is not a path. This is the main case for allowing unescaped
"$()" output, after all. Users might set PKG_CONFIG_TOP_BUILD_DIR to a
string with meta characters, but it no longer has the desired effect.
It's an unfortunate design flaw in the original pkg-config that the
contents of this variable are ambiguous. Either it's a path or it's a
piece of script. In practice it appears that, when set, the contents are
a raw path, without escaping, not a piece of script.

As a side effect, the definition of white space has shrunk to just four
characters, dropping backspace (\b) and form feed (\f). I suspect that
backspace was a fat-finger typo in the original pkg-config source and
that vertical tab (\v) was intended. Since nobody ever noticed or cared
about the mistake in the original pkg-config, it seems safe to drop this
bug-matching behavior.