Add manual page
This patch adds a small u-config.1 manual page, formatted using the
mdoc(7) language. In order to avoid having to keep up to date both the
manual page and the u-config.c source, the manual page defers to the
`--help` option when it comes to which pkg-config options and
environment variables are supported.

Debian has a policy of always shipping manual pages with binaries in
/usr/bin, hence why I wrote this.