Cast GENERIC_READ to signed 32-bit
Its value is in range of u32 but not i32. Since the enum contains
negative values, it forces the enum to i64 in order to encompass all
values. It is then, in a sense, narrowing to cast GENERIC_READ to i32.
This is harmless, but I want to avoid implicit narrowing conversions.

MSVC has a long-standing bug incorrectly defining this as a 4-byte enum,
and Clang when built with MSVC compatability matches this bug. Otherwise
it defines an 8-byte enum, tripping -Wconstant-conversion (enabled by
default). GCC does not warn about narrowing because the constant still
fits in a 32-bit integer, and so the conversion falls under the more
general -Wsign-conversion, which I disable.

This -Wconstant-conversion case is redundant with -Wsign-conversion, and
it seem wrong that -Wno-sign-conversion still produces sign conversion
warnings through this overlapping warning. Though in this particular
case it was beneficial.