fix undefined pointer arithmetic
while the previous code worked, it was undefined to do pointer
arithmetic and have the result be outside of the object (in this
case `p->name`) or one past it.

do the arithmetic on byte pointer - which is allowed to alias
other types and thus should (i think) make this legal as the
result is still within the `pkg` struct.