Don't recurse with modversion
With recursive processing the output with --modversion flag may be
incorrectly ordered as can be seen in the added test case. Recursive
processing is also inconsistent with both pkg-config and pkgconf: if
a package declares a version and requires a nonexistent package then
both will print the version with the flag, while u-config will fail
complaining that the transitively required package is missing.