Emit processed packages in topological order
Fixes #6. By prepending a pkg to the list when it's popped off the stack
we ensure that all of its dependencies are already processed, which
means that the flags will be printed in the correct order.

This has the consequence that flags in the output are in reversed
order compared to the input packages, i.e. the invocation

```
  pkg-config --libs a b
```

will result in

```
  -lb -la
```

assuming that there are no depenencies between liba and libb.