Return pkgs from process
This commit is just a cosmetic one; given that `process` now produces
the `pkgs` struct in a single invocation I've made it responsible for
its creation. Since only the `head` field is used later (the `count`
usage can be replaced with check whether `head` is null) we could
consider returning it directly.