wasm: Implement setjmp/longjmp for instance reuse
Requires at least Clang 19.

Involves a bit toolchain magic, and Clang hard crashed several times
working it out, but it actually works. Also WABT can no longer read the
WASM binary. Clearly it's still the early days of WASM tooling, which
has yet to catch up with 1995, let alone 2025.

I'd prefer to implement setjmp in C, but Clang insists on importing the
__wasm_* functions. So a module may not supply its own definitions. Por
què? Such details are undocumented by LLVM. Regardless, this may be the
first Javascript implementation of WASM setjmp. I'd be happier with the
old __builtin_{set,long}jmp, but it's unimplemented for WASM, likely
indefinitely.

I might revert, not even restoring "abort" but trap to get out fast, as
the instance is trashed either way. In the meantime this is a durable
place to capture lessons learned.

https://github.com/WebAssembly/tool-conventions/blob/main/SetjmpLongjmp.md