wasm: Improve path handling
Allow input directory paths to come with or without trailing slashes.
Typically these inputs do not have trailing slashes, but with WASI it
must be able to express the root path as "/" i.e. with a trailing slash.
These changes are incomplete, in part because the WASI spec itself is
incomplete, and it's unclear how various edge cases are supposed to be
handled (e.g. repeated slashes).