Use documented `InMemoryOrderModuleList`
This member of the loader structures has been documented since the
Windows 7 SDK release and this will most likely never change.

The advantage of using this instead is that the first module is always
the process itself according to Windows Internals 7th Edition
(https://stackoverflow.com/a/77011133).

This avoids needing a lock and extra ntdll exports, but requires using a
"containerof" macro.