WIP Add $PKG_CONFIG_SYSROOT_DIR for cross builds
This implementation behaves differently from pkg-config and pkgconf, and
by doing so better supports the intended purpose of this feature. It is
closer in behavior to pkg-config.

For example, suppose we are cross-compiling a program, foo-cli, that
depends on libfoo, which places its run-time data at some given, fixed
path under ${prefix}/share:

    $ make -C libfoo install
    cc -c lib.c
    cc -shared -o libfoo.so lib.o
    cp libfoo.so /usr/lib/
    cp share/libfoo.bin /usr/share/
    cp libfoo.pc /usr/lib/pkgconfig/

    $ make -C foo-cli install
    cc -c -DFOO_DATADIR=/usr/share -I/usr/include main.c
    cc -L/usr/lib -o foo-cli main.o -lfoo
    cp foo-cli /usr/bin/

When cross-compiling we might stage this in a DESTDIR for packaging,
and then build foo-cli against the staged libfoo:

    $ make -C libfoo install DESTDIR=/build
    cross-cc -c lib.c
    cross-cc -shared -o libfoo.so lib.o
    cp libfoo.so /build/usr/lib/
    cp share/libfoo.bin /build/usr/share/
    cp libfoo.pc /build/usr/lib/pkgconfig/

    $ make -C foo-cli install DESTDIR=/build
    cross-cc -c -DFOO_DATADIR=/usr/share -I/build/usr/include main.c
    cross-cc -L/build/usr/lib -o foo-cli main.o -lfoo
    cp foo-cli /build/usr/bin/

Notice that when building foo-cli both -I and -L needed to prepend the
DESTDIR so that the cross-toolchain could find the library. Also notice
that the data path is untouched because this is a run-time path, not a
build-time path. The DESTDIR will not exist at run time. A .pc file will
derive these paths from ${prefix}, but only -I and -L should be affected
when building against DESTDIR. The hard-coded data path should never be
affected by DESTDIR. A possible libfoo.pc:

    prefix = /usr
    Name: libfoo
    Description: The foo library
    Version: 1
    Cflags: -DFOO_DATADIR=${prefix}/share -I${prefix}/include
    Libs: -L${prefix}/lib -lfoo

So then with PKG_CONFIG_SYSROOT_DIR we ought to get:

    $ export PKG_CONFIG_PATH=/build/usr/lib/pkgconfig
    $ export PKG_CONFIG_SYSROOT_DIR=/build
    $ cross-pkg-config --cflags --libs libfoo
    -DFOO_DATADIR=/usr/share -I/build/usr/include -L/build/usr/lib -lfoo

When foo-cli runs later on its native system, the data will be in its
proper /usr/share, not under the /build DESTDIR back on the build host.

If foo-cli were to depend on a second library not installed in DESTDIR,
instead installed in, say, the cross toolchain sysroot, its pkg-config
configuration must be unaffected by PKG_CONFIG_SYSROOT_DIR because it is
not located in the DESTDIR and so does not require adjustment.